Refactor package structure of RALHandler
